python_engine/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
python-multipart>=0.0.9
websockets>=12.0
httpx>=0.27.0
chromadb>=0.4.22
sentence-transformers>=2.6.0
faster-whisper>=1.0.0
TTS>=0.22.0
systeminformation>=5.21.0
psutil>=5.9.0
numpy>=1.26.0
pdfplumber>=0.10.0
python-docx>=1.1.0
aiofiles>=23.2.0
google-generativeai>=0.8.0
keyring>=24.0.0